pyproject.toml
[tool.poetry]
name = "flask-scaffold"
version = "0.1.0"
description = "A scaffold to speed up launching a flask project"
authors = [
    "Weiliang Li <to.be.impressive@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.11"

# flask
APIFlask = "^2.1.1"
flask = "^3.0.0"

# celery
celery = "^5.4.0"

# server
gunicorn = "^21.2.0"
aiohttp = "^3.9.5"
aiohttp_wsgi = "^0.10.0"

# db
psycopg2 = "^2.9.9"
redis = "^5.0.4"
hiredis = "^2.0.0"

python-dotenv = "^1.0.1"
fakeredis = "^2.23.0"

[tool.poetry.dev-dependencies]
mypy = "^1.10.0"
ipython = "^8.24.0"
black = "^24.4.2"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
celery-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
profile = "black"
